+++ CHANGELOG.md
@@ -26,3 +26,6 @@
     heartbeat mechanism.
   - [Work in progress] Define the server/`Socket` class to represent clients
     connected to the server.
+- [Work in progress] Define the `EngineException` class:
+  - [Work in progress] Define the `ServerException` class.
+  - [Work in progress] Define the `TransportException` class.